FEAT: Update table definitions for act_log_admin, set_menu_admin, act_admin_notifications, and act_admin_fcm_tokens to ensure consistency in data types and indexing

+++ sql/2026-08-20-00005.sql
@@ -3,7 +3,7 @@
 CREATE TABLE `act_log_admin` (
 	`act_log_adminid` BIGINT NOT NULL,
 	`mst_adminid` INT NULL,
-	`set_menuid` INT NULL,
+	`set_menuid` SMALLINT NULL,
 	`action` VARCHAR(30) NULL,
 	`description` VARCHAR(200) NULL,
 	`ip_address` VARCHAR(200) NULL,
